% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centers.R
\docType{data}
\name{centers}
\alias{centers}
\title{Data to create hexmaps in R.}
\format{
A tibble with 51 rows and 4 variables:
\describe{
\item{staet.abb}{State abbreviation}
\item{x.center}{X value to plot label.}
\item{y.center}{Y value to plot label}
\item{id}{State name.}
}
}
\usage{
centers
}
\description{
Data contains the state abbreviations and points to plot them within the hexmap.
}
\keyword{datasets}
