% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ushexmapz-package.R
\docType{package}
\name{ushexmapz-package}
\alias{ushexmapz}
\alias{ushexmapz-package}
\title{ushexmapz: Easy Plotting for State Hexbin Maps}
\description{
This package is an easy way to plot tidy hexmaps without the need to directly handle geojsons, shapefiles, or other forms of geographic data. This allows users to spend more time making their maps look good and less time on loading geographic data into R.
}
\author{
\strong{Maintainer}: Jesse allagashanalytics@gmail.com Clark

}
\keyword{internal}
